Nested JS in submodules to be imported
